refactor(schema): rename environment and project files
